[fleche] Support workspace updates
We introduce a new coq/workspace_update notification, which tells
Flèche that .vo files in the workspace may have changed.

This will make coq-lsp pick up changes in the filesystem and restart
checking of Rocq's Requires when needed. The algorithm is not smart at
all yet, as it invalidates all the requires for all open files.

In this case, we set the document to a `WorkspaceUpdated` state, which
will force a full resumption of the document and the re-creation of
its initial state.