[petanque] New methods `petanque/proof_info` and `petanque/proof_info_at_pos`
The methods provide information about the current proof, if there is
one.

In particular, the `petanque/proof_info` method will return proof name
and original goal, as understood by the proof engine.

The `petanque/proof_info_at_pos` method will use document-level
information, so it will return the position of the statement and the
actual textual statement of the lemma.

Co-authored-by: DikieDick <26772815+DikieDick@users.noreply.github.com>