[rocq] Create output directory on .vo file save
This is very useful in the web context where we don't yet have a
proper virtual FS setup, thus we request to save a
`/working_dir/foo.vo` with the "Save .vo file" command, however it
fails due `/working_dir` not being present in the worker.

We should properly fix this by implementing the LSP virtual FS, but
for now this should do the trick.

In order for this to work well, we also modify the worker workspace
setup so we can add the LSP root folder to Rocq's search path.

cc: #833