[fleche] [rocq] New options and data fields to expose document comments
We introduce new options and data fields to expose document comments,
as understood by the Rocq lexer.

Due to lack of testing, this is an experimental feature; pass the
`--record_comments` flags to `rocq-lsp`, `fcc`, or `pet-server` to
enable it.

Once this flag is passed, comments for a document will be stored in
the new doc.comments field. Use the show_comments_on_hover option to
display the comments on hover for debugging purposes.

Co-authored-by: alexJ <37084688+blackbird1128@users.noreply.github.com>